[refactor] : 경매 추천 메서드 auction service로 이동
