+++ core/src/test/java/dev/handsup/auction/service/AuctionServiceTest.java
@@ -15,6 +15,9 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.SliceImpl;
+import org.springframework.data.domain.Sort;
 import org.springframework.test.util.ReflectionTestUtils;
 
 import dev.handsup.auction.domain.Auction;
@@ -24,9 +27,12 @@ import dev.handsup.auction.domain.product.ProductStatus;
 import dev.handsup.auction.domain.product.product_category.ProductCategory;
 import dev.handsup.auction.dto.request.RegisterAuctionRequest;
 import dev.handsup.auction.dto.response.AuctionDetailResponse;
+import dev.handsup.auction.dto.response.RecommendAuctionResponse;
 import dev.handsup.auction.exception.AuctionErrorCode;
+import dev.handsup.auction.repository.auction.AuctionQueryRepository;
 import dev.handsup.auction.repository.auction.AuctionRepository;
 import dev.handsup.auction.repository.product.ProductCategoryRepository;
+import dev.handsup.common.dto.PageResponse;
 import dev.handsup.common.exception.NotFoundException;
 import dev.handsup.fixture.AuctionFixture;
 import dev.handsup.fixture.ProductFixture;
@@ -42,6 +48,9 @@ class AuctionServiceTest {
 
 	@Mock
 	private AuctionRepository auctionRepository;
+
+	@Mock
+	private AuctionQueryRepository auctionQueryRepository;
 
 	@Mock
 	private ProductCategoryRepository productCategoryRepository;
@@ -112,5 +121,20 @@ class AuctionServiceTest {
 		assertThatThrownBy(() -> auctionService.getAuctionEntity(auctionId))
 			.isInstanceOf(NotFoundException.class)
 			.hasMessageContaining(AuctionErrorCode.NOT_FOUND_AUCTION.getMessage());
+	}
+
+	@DisplayName("정렬 조건에 따라 추천 경매를 조회할 수 있다.")
+	@Test
+	void getRecommendAuctions() {
+		//given
+		PageRequest pageRequest = PageRequest.of(0, 5, Sort.by("북마크수"));
+		ReflectionTestUtils.setField(auction, "createdAt",LocalDateTime.now());
+		given(auctionQueryRepository.sortAuctionByCriteria(null, null, null, pageRequest))
+			.willReturn(new SliceImpl<>(List.of(auction), pageRequest, false));
+		//when
+		PageResponse<RecommendAuctionResponse> response
+			= auctionService.getRecommendAuctions(null, null, null, pageRequest);
+		//then
+		assertThat(response.content().get(0)).isNotNull();
 	}
 }
